remuxer.sh
Got a bunch of help from GPT to get started on this. I know how it would work in concept, I didn't want to do something wrong trying to figure it out myself though, since this is something I want to be careful with. I feel more confident in having it set things up initially, then using my own safety checks to ensure that it's doing everything as intended. I think it would forget less simple mistakes then I probably would. I know it can hallucinate, but I am able to see that, sometimes it's not able to.

Prompt:

I have a bunch of songs I downloaded from Bandcamp over the last 4 years or so. I mistakenly have downloaded them in a lower-quality AAC encoding than I intended to. The rest of my library is downloaded from the iTunes Store, which means it follows a standard of at least AAC/256kbps. All of my music has custom metadata that I have edited to have complete artist info, artwork, etc.

I plan to re-download those incorrectly-encoded songs from Bandcamp, this time as lossless. I want to write a Bash script that uses `ffmpeg`, where it will preserve the same .m4a container as the original copy in my library (to persist my existing metadata tags), but simply replacing the media stream with the higher quality re-encode of that lossless file, remuxed to match the other files, which would be AAC/256kbps. I am only downloading the lossless copies because I want to use those as a source for the remux, I do not plan to include the lossless copies in the newly encoded files.

Code and response:

How it Works:
1. Extracts metadata from your original `.m4a` files.
2. Encodes the corresponding lossless file to AAC/256kbps using `libfdk_aac` (or `aac` if `libfdk_aac` isn't available).
3. Remuxes the new AAC stream into an `.m4a` container while applying the extracted metadata.
4. Cleans up any temporary files.

Requirements:
- ffmpeg (must be installed)
- libfdk_aac (preferred, but built-in `aac` codec can be used if unavailable)
- Ensure your lossless files are named identically to the original `.m4a` files, except for the extension.