fuzzy-rename.js
Prompt:

I have two folders of songs for a single album, one is named in one format, another is named in another format. They both have the same song name somewhere in the file name string. I want to write a JS script that will 'fuzzy search' the name of the first folder to that of the closest file name in the second folder. Essentially, at least one song will have the same name somewhere in it's file name format, and I want to find the correct one that does. Once that is the case, I want to rename that file to the exact name in folder 1. So I want all of the song files for the album in folder 2 to have the same names as the ones in folder 1. They do have different extensions though, so beware of that. It should persist the current extensions that the files in folder 2 has.