Skip Matching Names
Prompt:

Can you add a fix, to where it won't rename anything if it already exists in the other folder? Like if the name is already perfectly matching to that of over in folder one, it will skip it?

And it looks like it moved `files2Names` out of the loop scope for performance and calculation logic. Having it be set only once may make more sense. I want to debug that this is right though.