Preserve File Modification Dates
I wanted solely this originally, with the date added commits. GPT added the in-place edits on top of it though, so that's why I reverted here in the previous commit.

Still have yet to properly demo this out, I wanted to make sure I could easily test it, let alone before trying to use it.