afconvert iTunes Plus Format!!
Okay, I think we're on the right track here!!!

Doing some debugging and research as to what is the proper format for how I should encode these. I want it to either be exactly the same as the encoding you get when importing your own media to iTunes, or when downloaded from the store. So from the sound of it, it's AAC/256kbps CVBR from the store and VBR when importing from local files. VBR has the ability to be higher quality slightly, so I'm going to do that, in trade for a very smaller little bit of file size.

It looks like afconvert has more expected quality, while using ffmpeg with aac_at is less configurable, and you are only able to use CVBR, which is unfortunate.

These are the commands I demoed out to see the difference between the outputs of all of them. afconvert also has support for seamless playback I think, or something like that. This still works through the CLI too, which is great, as I ideally want to use this to generate things programatically rather than manually myself with one of the GUIs. The script here isn't completely working yet though, the main issue is that the metadata section isn't working. I need to refactor it a bit. The other option was the mastering tools droplet, I don't completely know the format that that is encoded with though, may have to debug it more beforehand. This command working consistently is the ideal scenario though, so I'm glad I have this.

Infinity is an absolutely outstanding album, I'm so glad I can hear it in this much more of high quality. Can't wait to do this to the rest of my library that was from Bandcamp 🥲 (joy)

```sh
ffmpeg -i ./1-01\ Truth.flac -vn -acodec aac_at -aac_at_mode 3 -aac_at_quality 2 ./1-01\ Truth.m4a
afconvert -f m4af -d aac -s 3 -ue vbrq 127 -q 127 ./1-01\ Truth.flac ./1-01\ Truth2.m4a
afconvert -f m4af -d aac -s 2 -b 256000 -q 127 ./1-01\ Truth.flac ./1-01\ Truth-cvbr.m4a
```

https://wiki.hydrogenaud.io/index.php?title=Apple_AAC
https://apple.stackexchange.com/questions/365616/what-are-the-command-line-options-for-afconvert
https://www.reddit.com/r/ffmpeg/comments/d9qokx/difference_between_aac_at_and_afconvert_encoders/
https://news.ycombinator.com/item?id=29680166
https://www.reddit.com/r/ffmpeg/comments/gqws7c/unable_to_convert_from_flac_to_aac_using_aac_at/
https://hydrogenaud.io/index.php/topic,70405.0.html
https://www.reddit.com/r/AppleMusic/comments/v5wgvt/what_happens_to_songs_encoded_using_the_itunes/
https://apple.stackexchange.com/questions/46076/what-are-the-afconvert-settings-for-the-itunes-plus-aac-encoding-setting
https://www.apple.com/apple-music/apple-digital-masters/
https://www.youtube.com/watch?v=yztZnCXFr7E
https://www.youtube.com/watch?v=B-bwfEIJGkQ
https://www.youtube.com/watch?v=F-FrPOW_Fn4
https://discussions.apple.com/thread/5862481?sortBy=rank
https://stackoverflow.com/questions/59549766/is-there-a-way-to-identify-if-an-audio-is-vbr-or-cbr-using-ffmpeg