Temp File Uncomment
This was there because I wanted to ensure/debug that the temp file was being written correctly in this script. It is! I also downloaded a 'problematic' AAC encode from Bandcamp again to compare how much this encode is to that. For this new album ("We All Hold It Up Together"), things are still twice as big at the file level (at minimum), and the FLAC is still much bigger than that, usually at minimum 2x to the nicely encoded AAC files we get with this script.