Ambience CD Reading
Found a CD at a store today, from 1994, and rather than being an audio CD, it stores AIFF files in an HFS file system.

Went through a whole hassle to open the files, as macOS doesn't support HFS file systems as of the last few updates.

Used my Ubuntu machine to open it, then copied the files back to the flash drive, and realized none of the files had extensions on them. These scripts rename the files to add the extensions, and I also made it so I could view all of the sounds all at once, since they are nested in a bunch of folders.

https://rarefaction.com/textural.html
https://www.rarefaction.com/soundindexes/TexturalAmbienceSoundFileIndex.pdf

Some of the files were corrupted or something it seems, as trying to copy them or open them, resulted in an input/output error. It seems to be the ones at the end of the disc, the last two Particle fission ones.

I decided to get it because the spine says "Textural Abmience", and I was curious if that was on purpose or not. I don't think it was. The front doesn't have the misspelling though, which is also interesting.