ALAC Bandcamp Downloads
I've been using this for the last few Bandcamp downloads of mine, committing it now though, now that I'm committed to it over FLAC downloads. I like that it has all of the metadata, and artwork, all packaged and ready to go. It's just the codec in the container that I want to remux to a nice AAC encode.