Qobuz Importer
Adding a dedicated FLAC importer, because up until now I have been shoehorning the one I use for M4A ALAC from Bandcamp. Now this will work the same for FLAC, without needing to edit the script to adjust things.

First buy with Qobuz, got the 2007 mix of Selling England by the Pound. It's awesome. Loving the sound of it.

Was debating getting the CD, but I moreso want the lossless quality in general, than specifically the hard copy. And, I got it today, can listen today. Otherwise, you'd have to wait for shipping, possibly a used disc, all of that kind of thing. The only thing here was that I had to research a bunch about these "HiRes lossless services". There's a whole bunch of "consumer jargon", which isn't real tech talk I feel. CD quality is more than great, no need for high res formats for listening. Just wanted to make sure that the site was credible, since their marketing is less than ideal (stair steps, "HiRes", etc.). But, it's CD quality masters, so that's all I ask for. I even found out Carpincho is on there! So it's even one I distribute to, and I had no idea 😂; I do think I remember reading that it was in the list of supported services when I started CDBaby, just hadn't really looked into it previously.

Was heavily debating the original release of this CD, but ended up feeling I like the listening experience of the new one, even though I strongly adore the dynamics of the original. I really do like the sound of the new one though, it's a very pleasing listen. Again, as I mentioned in another commit, all of this is subjective. That's why music is so messy, because it's inherently human.